+++ autd3/src/datagram/mod.rs
@@ -1,2 +1,3 @@
 pub mod gain;
 pub mod modulation;
+pub mod stm;